knoggin-server/pyproject.toml
[project]
name = "knoggin-server"
version = "0.1.0"
description = "A transparent, self-hosted knowledge graph for personal AI memory - Server Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "neo4j>=6.0.3",
    "numpy>=2.3.5",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "redis>=7.1.0",
    "redis-tools>=0.0.2",
    "sentence-transformers>=5.2.0",
    "transformers==4.41.2",
    "torch>=2.6.0,<2.7.0",
    "xformers>=0.0.29.post3",
    "scikit-learn>=1.8.0",
    "gliner>=0.2.22",
    "spacy>=3.8.11",
    "wordfreq>=3.1.1",
    "en-core-web-md",
    "huggingface-hub[cli]>=0.36.0",
    "chromadb>=1.4.1",
    "docx2txt>=0.9",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.6.2",
    "mcp>=1.2.0,<2.0",
    "duckduckgo-search>=8.0.0",
    "cachetools>=5.4.0",
    "markitdown>=0.1.5",
    "rank-bm25>=0.2.2",
    "ddgs>=9.11.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0"
]

[project.scripts]
knoggin = "cli.main:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cli*", "api*", "db*", "common*", "core*", "agent*", "jobs*", "scripts*", "log*"]

[tool.setuptools.package-data]
cli = ["templates/*"]

[tool.uvicorn]
reload-exclude = ["redis_data", "memgraph_data", "*.log"]

[tool.uv.sources]
en-core-web-md = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl" }

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
